Fixes npm version instructions
The instructions in our README file are still not correct for actually releasing updates and publishing to npm, and we're either publishing a new version without the changes or failing to push the tags appropriately.

These changes should address the main issue we're having, but I would like to look more closely about the `--tags` option on `git push` and why just the one step doesn't seem to work.  For the time being, two pushes (one for the change to `package.json` itself and another for the tag update) at least ensures that we get exactly what we're expecting.