fixing rem calc bug
I am fixing this in one place (where it was causing me an issue but it may be a widespread problem I believe. Anyplace the a `calc(X - Yrem)` is used in order to offset by an element of width `u(Yrem)`, the offset will be broken when you adapt the mode (eg,. fec-eregs sets $px-only to `true`, which surfaced this bug).