Mitigitate XSS and content spoofing attacks
This changeset adds support to mitigate potential XSS and content spoofing attacks through our client-side code.  Penetration testing has yielded a few spots in our code where we were accepting values as is when pulling them from the query string in the URL, mainly when pre-initializing filters.  We have addressed this by making use of the bleach (https://github.com/ecto/bleach) library to strip all tags from this content, and then stripping out all other characters not whitelisted as being valid.  We determined that only certain characters should be allowed for the filter values and feel safe in ignoring all others.

A huge thank you to @xtine for pairing with me on this and helping solve the issue!  h/t to @LinuxBozo, @toolness, and @cmc333333 for the additional help!