Fix issue related to jQuery version bump
The update to jQuery caused a regression to appear with one of our selectors, though the selector itself was not set up quite right.  This changeset fixes the issue by adding quotes around the value for the `a[ref="<value>"]` selector.  Please see the comments section of the version 1.12 and 2.2 release notes for details: http://blog.jquery.com/2016/01/08/jquery-2-2-and-1-12-released/