swap out references to beta from README
This replaces `betaFEC` with `FEC.gov`, `the new FEC.gov`, or `the FEC` depending on situation/context.

One additional question: I've noticed a handful of the repos list five repos under `All repositories`. But a couple list more than that. I am assuming this is primarily about when the repo was created and these are things that ought to be reconciled? Or is there more to the choice than that?